Changed servlets so that one servlet, getElection, fetches election information from the Civic Information API and puts it in Datastore, while the second servlet, election, retrieves that information from Datastore. /getElection is called by the cron job every night, while /election is called within our code whenever the data from Datastore is needed. Also now use the Secrets API to hide the API key for the project once it's deployed, and include option to use local environment variable during testing on local server.
